RDRP-780: separate validate postcodes
