+++ src/developer_config.yaml
@@ -7,21 +7,21 @@ global:
   dev_test : False
   network_or_hdfs: network #whether to load from hdfs or network (local Python)
   # Staging and validation settings
-  postcode_csv_check: False
+  postcode_csv_check: True
   load_updated_snapshot: False # Whether to load the updated snapshots for amendments and additions
   load_ni_data: True
   load_historic_data: False
-  run_construction: False
+  run_construction: True
   run_ni_construction: False
   load_manual_outliers: False
   load_manual_imputation: False
   load_backdata: True  # whether to load previous year data for MoR
   load_reference_list: True # the reference list corrections should always be loaded in year 1
-  load_from_feather: True
+  load_from_feather: False
   # Apportionment to sites settings
   apportion_sites: True
   # Output settings
-  output_full_responses: False
+  output_full_responses: True
   output_ni_full_responses: False
   output_imputation_qa: False
   output_auto_outliers: False
@@ -32,7 +32,7 @@ global:
   output_short_form: False
   output_gb_sas: False
   output_ni_sas: False
-  output_tau: False
+  output_tau: True
   output_intram_by_pg_gb: False
   output_intram_by_pg_uk: False
   output_intram_by_itl1: False
