+++ src/imputation/MoR.py
@@ -4,7 +4,7 @@ import pandas as pd
 import numpy as np
 import re
 
-from src.staging.validation import postcode_topup
+from src.staging import postcode_validation as pcval
 from src.imputation.apportionment import run_apportionment
 from src.imputation.tmi_imputation import (
     create_imp_class_col,
@@ -133,9 +133,9 @@ def carry_forwards(df, backdata, impute_vars):
     # Copy values from relevant columns where references match
     match_cond = df["_merge"] == "both"
 
-    # Apply the postcode top-up to clean the postcodes in col 601 of the back data
+    # Apply the postcode formatting to clean the postcodes in col 601 of the back data
     df.loc[match_cond, "601_prev"] = df.loc[match_cond, "601_prev"].apply(
-        postcode_topup
+        pcval.format_postcodes
     )
 
     # Replace the values of certain columns with the values from the back data
@@ -145,7 +145,7 @@ def carry_forwards(df, backdata, impute_vars):
 
     # Update the postcodes_harmonised column from the updated column 601
     df.loc[match_cond, "postcodes_harmonised"] = df.loc[match_cond, "601"]
-    
+
     # Update the varibles to be imputed by the corresponding previous values
     for var in impute_vars:
         df.loc[match_cond, f"{var}_imputed"] = df.loc[match_cond, f"{var}_prev"]
@@ -249,7 +249,7 @@ def get_threshold_value(config: dict) -> int:
     else:
         raise Exception("The variable 'mor_threshold' in the 'imputation' section "
                         "of the config must be zero or a positive integer.")
-        
+
 
 def group_calc_link(group, target_vars, config):
     """Apply the MoR method to each group
@@ -284,7 +284,7 @@ def group_calc_link(group, target_vars, config):
                 valid_group_size = False
 
         # If the group is a valid size, and there are non-null, non-zero values for this
-        # 'var', then calculate the mean        
+        # 'var', then calculate the mean
         if valid_group_size & (sum(~group[f"{var}_gr_trim"] & non_null_mask) != 0):
             group[f"{var}_link"] = group.loc[
                 ~group[f"{var}_gr_trim"] & non_null_mask, f"{var}_gr"
