+++ src/staging/staging_helpers.py
@@ -11,6 +11,7 @@ from typing import Callable, Tuple, Dict
 # Our own modules
 from src.utils.wrappers import time_logger_wrap
 from src.staging import validation as val
+from src.staging import postcode_validation as pcval
 from src.staging import spp_parser, history_loader
 from src.staging import spp_snapshot_processing as processing
 
@@ -455,7 +456,7 @@ def stage_validate_harmonise_postcodes(
     postcode_masterlist = postcode_mapper["pcd2"]
 
     # Validate the postcode column in the full_responses DataFrame
-    invalid_df = val.validate_post_col(full_responses, postcode_masterlist, config)
+    invalid_df = pcval.run_full_postcode_process(full_responses, postcode_masterlist, config)
 
     # Log the saving of invalid postcodes to a file
     StagingHelperLogger.info("Saving Invalid Postcodes to File")
