+++ src/staging/validation.py
@@ -1,6 +1,5 @@
 import os
 import toml
-import postcodes_uk
 import pandas as pd
 import numpy as np
 
@@ -9,361 +8,6 @@ from src.utils.wrappers import time_logger_wrap, exception_wrap
 
 # Set up logging
 ValidationLogger = logging.getLogger(__name__)
-
-
-def postcode_topup(mystr: str, target_len: int = 8) -> str:
-    """
-    Regulates the number of spaces between the first and the second part of
-        a postcode, so that the total length is 8 characters.
-    Brings all letters to upper case, in line with the mapper.
-    Splits the postcode string into parts, separated by any number of spaces.
-    If there are two or more parts, the first two parts are used.
-    The third and following parts, if present, are ignored.
-    Calculates how many spaces are needed so that the total length is 8.
-    If the total length of part 1 and part 2 is already 8, no space will be
-        inserted.
-    If their total length is more than 8, joins part 1 and part 2 without
-        spaces and cuts the tail on the right.
-    If there is only one part, keeps the first 8 characters and tops it up with
-        spaces on the right if needed.
-    Empty input string would have zero parts and will return an empty string.
-
-    Args:
-        mystr (str): Input postcode.
-        target_len (int): The desired length of the postcode after topping up.
-
-    Returns:
-        str: The postcode topped up to the desired number of characters.
-    """
-    if pd.notna(mystr):
-        mystr = mystr.upper()
-        parts = mystr.split()
-
-        if len(parts) == 1:
-            mystr = mystr.strip()
-            part1 = mystr[:-3]
-            part2 = mystr[-3:]
-
-            num_spaces = target_len - len(part1) - len(part2)
-            if num_spaces >= 0:
-                return part1 + " " * num_spaces + part2
-            else:
-                return (part1 + part2)[:target_len]
-
-        elif len(parts) >= 2:
-            part1 = parts[0]
-            part2 = parts[1]
-
-            num_spaces = target_len - len(part1) - len(part2)
-            if num_spaces >= 0:
-                return part1 + " " * num_spaces + part2
-            else:
-                return (part1 + part2)[:target_len]
-
-        else:
-            return ""
-
-
-class ColumnMismatch(Exception):
-    def __init__(self, message=None):
-        self.message = message
-        super().__init__(message)
-
-
-def validate_postcode_pattern(pcode: str) -> bool:
-    """A function to validate UK postcodes which uses the postcodes_uk package
-    to verify the pattern of a postcode by using regex.
-
-    Args:
-        pcode (str): The postcode to validate
-
-    Returns:
-        bool: True or False depending on if it is valid or not
-    """
-
-    if pcode is None:
-        return False
-
-    pcode = pcode.upper().strip()
-
-    # Validation step
-    valid_bool = postcodes_uk.validate(pcode)
-
-    return valid_bool
-
-
-@time_logger_wrap
-@exception_wrap
-def validate_post_col(
-    df: pd.DataFrame, postcode_masterlist: pd.DataFrame, config: dict
-):
-    """This function checks if all postcodes in the specified DataFrame column
-        are valid UK postcodes. It uses the `validate_postcode_pattern` function to
-        perform the validation.
-
-        - A `postcodes_harmonised` column is created using the value in col `601`,
-        or the value in col `postcodereference` (IDBR) where col `601` is blank.
-        - Validation checks run on `postcodes_harmonised`
-        First we validate the pattern of the postcode.
-        Secondly if a postcode is valid, we validate that the postcodes are real
-        - One dataframe containing any invalid postcodes outputted with relevant
-        record information (inc. postcode source).
-        - Any invalid postcodes removed from the `postcodes_harmonised`
-        column in dataframe.
-
-    Args:
-        df (pd.DataFrame): The DataFrame containing the postcodes.
-        postcode_masterlist (pd.DataFrame): The dataframe containing the correct
-        postocdes to check against
-        config (dict): The postcode settings from the config settings
-
-    Returns:
-        combined_invalid_postcodes_df (pd.DataFrame): A dataframe of invalid postcodes,
-        either with the incorrect pattern or not found in the masterlist
-    """
-    if not isinstance(df, pd.DataFrame):
-        raise TypeError(f"The dataframe you are attempting to validate is {type(df)}")
-
-    # Create new column and fill with "601" and the nulls with "referencepostcode"
-    df["postcodes_harmonised"] = df["601"].fillna(df["referencepostcode"])
-
-    # Create a copy to work from and add temp "postcode_source" column
-    validation_df = df.copy()
-    validation_df["postcode_source"] = validation_df.apply(
-        lambda row: "column '601'"
-        if pd.notna(row["601"])
-        else "column 'referencepostcode' (IDBR)",
-        axis=1,
-    )
-
-    # Apply the pattern validation function
-    invalid_pattern_postcodes = validation_df.loc[
-        ~validation_df["postcodes_harmonised"].apply(validate_postcode_pattern),
-        "postcodes_harmonised",
-    ]
-
-    # Save to df
-    invalid_df = pd.DataFrame(
-        {
-            "reference": validation_df.loc[
-                invalid_pattern_postcodes.index, "reference"
-            ],
-            "instance": validation_df.loc[invalid_pattern_postcodes.index, "instance"],
-            "formtype": validation_df.loc[invalid_pattern_postcodes.index, "formtype"],
-            "postcode_issue": "invalid pattern / format",
-            "incorrect_postcode": invalid_pattern_postcodes,
-            "postcode_source": validation_df.loc[
-                invalid_pattern_postcodes.index, "postcode_source"
-            ],
-        }
-    )
-
-    # Log the invalid postcodes
-    ValidationLogger.warning(
-        f"Invalid pattern postcodes found: {invalid_pattern_postcodes.to_list()}"
-    )
-    ValidationLogger.warning(
-        f"""Number of invalid pattern postcodes found:
-        {len(invalid_pattern_postcodes.to_list())}"""
-    )
-
-    # Remove the invalid pattern postcodes before checking if they are real
-    val_df = validation_df.loc[
-        ~validation_df.index.isin(invalid_pattern_postcodes.index.to_list())
-    ]
-
-    # Clean postcodes to match the masterlist
-    check_real_df = clean_postcodes(val_df, "postcodes_harmonised")
-
-    # Only validate not null postcodes for the column "601"
-    check_real_df = check_real_df.loc[~check_real_df["postcodes_harmonised"].isnull()]
-
-    # Create a list of postcodes not found in masterlist in col "postcodes_harmonised"
-    unreal_postcodes = check_pcs_real(
-        val_df, check_real_df, postcode_masterlist, config
-    )
-
-    # Save to df
-    unreal_df = pd.DataFrame(
-        {
-            "reference": validation_df.loc[unreal_postcodes.index, "reference"],
-            "instance": validation_df.loc[unreal_postcodes.index, "instance"],
-            "formtype": validation_df.loc[unreal_postcodes.index, "formtype"],
-            "postcode_issue": "not found in masterlist",
-            "incorrect_postcode": unreal_postcodes,
-            "postcode_source": validation_df.loc[
-                unreal_postcodes.index, "postcode_source"
-            ],
-        }
-    )
-
-    # Log the unreal postcodes
-    ValidationLogger.warning(
-        f"These postcodes are not found in the ONS postcode list: {unreal_postcodes.to_list()}"  # noqa
-    )
-
-    ValidationLogger.warning(
-        f"Number of postcodes not found in the ONS postcode list: {len(unreal_postcodes.to_list())}"  # noqa
-    )
-
-    # Combine the two lists for logging
-    combined_invalid_postcodes = pd.concat(
-        [unreal_postcodes, invalid_pattern_postcodes]
-    )
-    combined_invalid_postcodes.drop_duplicates(inplace=True)
-
-    if not combined_invalid_postcodes.empty:
-        ValidationLogger.warning(
-            f"Total list of unique invalid postcodes found: {combined_invalid_postcodes.to_list()}"  # noqa
-        )
-
-        ValidationLogger.warning(
-            f"Total count of unique invalid postcodes found: {len(combined_invalid_postcodes.to_list())}"  # noqa
-        )
-
-    # Combine and sort the two dataframes for output
-    combined_invalid_postcodes_df = pd.concat([invalid_df, unreal_df])
-
-    combined_invalid_postcodes_df = combined_invalid_postcodes_df.sort_values(
-        ["reference", "instance"], ascending=[True, True]
-    ).reset_index(drop=True)
-
-    # update df to exclude any invalid postcode entries
-    df["postcodes_harmonised"] = df["postcodes_harmonised"].where(
-        ~df["postcodes_harmonised"].isin(
-            combined_invalid_postcodes_df["incorrect_postcode"]
-        ),
-        other=None,
-    )
-
-    df["postcodes_harmonised"] = df["postcodes_harmonised"].apply(postcode_topup)
-    df["601"] = df["601"].apply(postcode_topup)
-
-    ValidationLogger.info("All postcodes validated....")
-
-    return combined_invalid_postcodes_df
-
-
-def insert_space(postcode, position):
-    """Automate the insertion of a space in a string"""
-    return postcode[0:position] + " " + postcode[position:]
-
-
-@exception_wrap
-def get_masterlist(postcode_masterlist) -> pd.Series:
-    """This function converts the masterlist dataframe to a Pandas series
-
-    Returns:
-        pd.Series: A series of postcodes
-    """
-    masterlist = postcode_masterlist.squeeze()
-
-    return masterlist
-
-
-def clean_postcodes(df, column):
-    """Clean the postcodes to fit the masterlist format
-    This is done by applying all the following steps:
-
-    - remove whitespaces on postcodes of 8+ characters
-    - add 1 whitespace for 7 digit postcodes
-    - add 2 whitespaces for 6 digit postcodes
-    - add 3 whitespaces for 5 digit postcodes
-
-    Once all the postcodes are 8 characters long, they match the pcd2 column
-    we is used to validate postcodes.
-    """
-
-    # Create a copy df to manipulate
-    check_real_df = df.copy()
-
-    # Convert to uppercase and strip whitespaces at start/end
-    check_real_df[column] = check_real_df[column].str.upper()
-    check_real_df[column] = check_real_df[column].str.strip()
-
-    # Renove whitespaces on larger than 8 characters
-    check_real_df.loc[
-        check_real_df[column].str.strip().str.len() > 8, column
-    ] = check_real_df.loc[
-        check_real_df[column].str.strip().str.len() > 8, column
-    ].str.replace(
-        " ", ""
-    )
-
-    # Add whitespace to 7 digit postcodes
-    check_real_df.loc[
-        check_real_df[column].str.strip().str.len() == 7, column
-    ] = check_real_df.loc[
-        check_real_df[column].str.strip().str.len() == 7, column
-    ].apply(
-        lambda x: insert_space(x, 4)
-    )
-
-    # Add 2 whitespaces to 6 digit postcodes
-    check_real_df.loc[check_real_df[column].str.strip().str.len() == 6, column] = (
-        check_real_df.loc[check_real_df[column].str.strip().str.len() == 6, column]
-        .apply(lambda x: insert_space(x, 3))
-        .apply(lambda x: insert_space(x, 4))
-    )
-
-    # Add 3 whitespaces to 5 digit postcodes
-    check_real_df.loc[check_real_df[column].str.strip().str.len() == 5, column] = (
-        check_real_df.loc[check_real_df[column].str.strip().str.len() == 5, column]
-        .apply(lambda x: insert_space(x, 2))
-        .apply(lambda x: insert_space(x, 3))
-        .apply(lambda x: insert_space(x, 4))
-    )
-
-    return check_real_df
-
-
-def check_pcs_real(
-    df: pd.DataFrame,
-    check_real_df: pd.DataFrame,
-    postcode_masterlist: pd.DataFrame,
-    config: dict,
-):
-    """Checks if the postcodes are real against a masterlist of actual postcodes.
-
-    In the masterlist, all postcodes are 7 characters long, therefore the
-    reference are formatted to match this format.
-
-    All postcodes above 7 characters are stripped of whitespaces.
-    All postcodes less than 7 characters have added whitespaces in the middle.
-
-    This formatting is applied to a copy dataframe so the original is unchanged.
-
-    The final output are the postcodes from the original dataframe
-
-    Args:
-        df (pd.DataFrame): The DataFrame containing the postcodes.
-        check_real_df (pd.DataFrame): The DataFrame excluding invalid postcodes, to
-        run the comparison against
-        postcode_masterlist (pd.DataFrame): The dataframe containing the correct
-        postocdes to check against
-
-    Returns:
-        unreal_postcodes (pd.DataFrame): A dataframe containing all the
-        original postcodes not found in the masterlist
-
-    """
-
-    if config["global"]["postcode_csv_check"]:
-        master_series = get_masterlist(postcode_masterlist)
-
-        # Check if postcode are real
-        check = check_real_df[
-            ~check_real_df["postcodes_harmonised"].isin(master_series)
-        ]
-        unreal_postcodes = df.loc[check.index, "postcodes_harmonised"]
-
-    else:
-        emptydf = pd.DataFrame(columns=["postcodes_harmonised"])
-        unreal_postcodes = emptydf.loc[
-            ~emptydf["postcodes_harmonised"], "postcodes_harmonised"
-        ]
-
-    return unreal_postcodes
 
 
 @exception_wrap
